card, hand, and deck classes done
I created the Hand class that consists of 3 attributes (Rank, Suit, and Value), with a constructor and getters for these attributes.

I created the deck class, with only 1 attribute which is the deck array list of card objects.  One method initDeck, will created every possible combo of cards, and the other method, shuffleDeck, will shuffle the order that these objects appear.

The Hand class consists of of three methods, drawCard, which draws a card from the shuffled deck, discardCard will discard a card from from the players hand, and getValue calculates the value of the players hand.