Update Hand.java
Changed the Hand attributes to of objects of the Card class as an array list.  

Cleaned up the drawCard method, by also changing the return type to be an array list of Card objects, as well as the input parameter being changed to an array list of Card objects as well.  Also, added a reference to the instance variable hand.  

Also cleaned up the discardCard method by changing the return value to be an array list of Card objects as well.  
Also, added a reference to the instance variable hand. 

Finally, I cleaned up the the getValue method by again adding a reference to the instance variable hand. 
Then I got rid of the giant if statement check for the cards value, I found an easier way of just using the getValue() method.  Also, added the cumulative sum within the for loop.