In the MainMenu class, I changed the name of the method showMainMenu to initializeMainMenu. I then use the constructor to call the initializeMainMenu method. I also changed the constructor to accept the player's username, so each unique player will get their own game of Blackjack started. In the InitializeMainMenu method, I put all the frame setup in it, as well as the panel setup, and the button initializations. I also added a method called buttonPresses in the MainMenu class that will listen for one of the three buttons to be pressed, i.e., Play Blackjack, Leaderboard, or Quit. Finally, I added the startGame method into the MainMenu class, for easier transition into gameplay and game creation after the Play Blackjack button is pressed.
In the Main file, I re-implemented the startMainMenu method, so that it's only called when a player successfully logs in or signs up.  This method then creates an instance of the MainMenu class, where we listen for button presses to know what to do, i.e., start the game, open the leaderboard, or quit.